// web/src/api/generated/model/loanOut.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TCFL Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface LoanOut {
  id: string;
  title: string;
  authors: string[];
  edition: string | null;
  due_at: string;
  renewals_left: number;
  renewals_max: number;
  overdue: boolean;
}